requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.7.0
filelock==3.13.1
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.40
huggingface-hub==0.20.1
idna==3.6
Jinja2==3.1.2
kiwisolver==1.4.5
kornia==0.7.1
lightning-utilities==0.10.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.4
natsort==8.4.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open-clip-torch==2.23.0
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
pipreqs==0.4.13
protobuf==4.25.1
psutil==5.9.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.1.3
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scipy==1.11.4
sentencepiece==0.1.99
sentry-sdk==1.39.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
timm==0.9.12
tokenizers==0.15.0
torch==2.1.2
torchaudio==2.1.2
torchmetrics==1.2.1
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
triton==2.1.0
typing_extensions==4.9.0
urllib3==2.1.0
wandb==0.16.1
wcwidth==0.2.12
xformers==0.0.23.post1
yarg==0.1.9
yarl==1.9.4
